localize non-hardware variables
Declare setFltrS, byp, vol, mix, and mode as static variables within AudioCallback(). The LoopChannels cannot be handled as local statics without an audio glitch occurring on the occasional startup. Instead they have to remain global variables, along with the hardware-related variables.
Since it is not possible to define a callable callback with extra parameters, and the loop needs to be cleared before entering the callback, they will remain global.